move noxfile.py to .github/scripts
