+++ .github/workflows/platform-core.yml
@@ -43,4 +43,4 @@ jobs:
       - name: Run tests
         run: |
           pip install nox
-          nox -f openbb_platform/noxfile.py -s tests --python ${{ matrix.python_version }}
+          nox -f .github/scripts/noxfile.py -s tests --python ${{ matrix.python_version }}
